設定周りの処理はstoreImplに切り出す
